app/requirements.txt
ttensorflow
pillow
fastapi
uvicorn
opencv-python-headless
gdown